# ENH: Improve performance for arrow dtypes in monotonic join