@phofl 2023-02-13
Arrays have 2 million entries, initial performance 380ms, on this pr

```
%timeit idx.union(idx2)
11.5 ms ± 477 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

I totally agree with your point regarding EA interface, this is a short term solution for 2.0